# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Real Estate Investment API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /withdrawal/request:
    post:
      operationId: requestWithdrawal
      tags: [withdrawal]
      summary: Submit a withdrawal request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WithdrawalRequestBody"
      responses:
        "201":
          description: Request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WithdrawalRequestItem"
...
components:
  schemas:
...
    WithdrawalRequestBody:
      type: object
      required: [amount, transactionPin]
      properties:
        amount:
          type: number
        transactionPin:
          type: string
    WithdrawalRequestItem:
      type: object
      required: [id, userId, amount, bankName, accountNumber, accountHolderName, status, adminNote, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        amount:
          type: number
        bankName:
          type: string
        accountNumber:
          type: string
        accountHolderName:
          type: string
        status:
          type: string
        adminNote:
          type: string
        createdAt:
          type: string
...